# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PayTrace Backend",
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/v1/incidents": {
      "post": {
        "tags": [
          "incidents",
          "incidents"
        ],
        "summary": "Create Incident",
        "operationId": "create_incident_api_v1_incidents_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IncidentCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "incidents",
          "incidents"
        ],
        "summary": "List Incidents",
        "operationId": "list_incidents_api_v1_incidents_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "status_filter",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status Filter"
            }
          },
          {
            "name": "scenario_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scenario Id"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}": {
      "get": {
        "tags": [
          "incidents",
          "incidents"
        ],
        "summary": "Get Incident",
        "operationId": "get_incident_api_v1_incidents__incident_id__get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/simulated": {
      "post": {
        "tags": [
          "incidents",
          "incidents"
        ],
        "summary": "Create Simulated Incident",
        "description": "Materialise a deterministic harness scenario and create its Incident.",
        "operationId": "create_simulated_incident_api_v1_incidents_simulated_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SimulatedIncidentCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IncidentResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/funnel": {
      "get": {
        "tags": [
          "incidents",
          "incidents"
        ],
        "summary": "Get Incident Funnel",
        "operationId": "get_incident_funnel_api_v1_incidents__incident_id__funnel_get",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Incident Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FunnelResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/diagnosis-runs": {
      "post": {
        "tags": [
          "incidents",
          "incidents"
        ],
        "summary": "Trigger Diagnosis",
        "description": "Trigger an async diagnosis run for an incident.\n\nRequires ``Idempotency-Key`` header per plan \u00a7 17.1.  Returns 202\nwith the DiagnosisRun ID immediately; the actual work happens in a\nCelery worker.",
        "operationId": "trigger_diagnosis_api_v1_incidents__incident_id__diagnosis_runs_post",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Incident Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisRunTriggerResponse"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/incidents/{incident_id}/diagnosis-runs/{run_id}/retry": {
      "post": {
        "tags": [
          "incidents",
          "incidents"
        ],
        "summary": "Retry Diagnosis",
        "description": "Retry a failed diagnosis run.\n\nCreates a new DiagnosisRun with a new idempotency key and dispatches\na fresh Celery task.  The original run is left as-is.",
        "operationId": "retry_diagnosis_api_v1_incidents__incident_id__diagnosis_runs__run_id__retry_post",
        "parameters": [
          {
            "name": "incident_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Incident Id"
            }
          },
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Idempotency-Key"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiagnosisRunRetryResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FunnelResult": {
        "properties": {
          "baseline": {
            "$ref": "#/components/schemas/FunnelPeriodStats"
          },
          "incident": {
            "$ref": "#/components/schemas/FunnelPeriodStats"
          },
          "deltas": {
            "items": {
              "$ref": "#/components/schemas/FunnelStageDelta"
            },
            "type": "array",
            "title": "Deltas"
          },
          "anomalous_stages": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Anomalous Stages"
          }
        },
        "type": "object",
        "required": [
          "baseline",
          "incident",
          "deltas",
          "anomalous_stages"
        ],
        "title": "FunnelResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IncidentCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "scenario_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Scenario Id"
          },
          "dataset_ref": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Dataset Ref"
          },
          "baseline_start": {
            "type": "string",
            "format": "date-time",
            "title": "Baseline Start"
          },
          "baseline_end": {
            "type": "string",
            "format": "date-time",
            "title": "Baseline End"
          },
          "incident_start": {
            "type": "string",
            "format": "date-time",
            "title": "Incident Start"
          },
          "incident_end": {
            "type": "string",
            "format": "date-time",
            "title": "Incident End"
          },
          "trigger_metric": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Trigger Metric"
          },
          "baseline_value": {
            "type": "number",
            "title": "Baseline Value"
          },
          "observed_value": {
            "type": "number",
            "title": "Observed Value"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "ontology_version": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Ontology Version"
          }
        },
        "type": "object",
        "required": [
          "title",
          "scenario_id",
          "dataset_ref",
          "baseline_start",
          "baseline_end",
          "incident_start",
          "incident_end",
          "trigger_metric",
          "baseline_value",
          "observed_value",
          "ontology_version"
        ],
        "title": "IncidentCreate",
        "description": "POST /incidents request body."
      },
      "IncidentListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/IncidentResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationMeta"
          }
        },
        "type": "object",
        "required": [
          "items",
          "pagination"
        ],
        "title": "IncidentListResponse"
      },
      "IncidentResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "scenario_id": {
            "type": "string",
            "title": "Scenario Id"
          },
          "dataset_ref": {
            "type": "string",
            "title": "Dataset Ref"
          },
          "baseline_start": {
            "type": "string",
            "format": "date-time",
            "title": "Baseline Start"
          },
          "baseline_end": {
            "type": "string",
            "format": "date-time",
            "title": "Baseline End"
          },
          "incident_start": {
            "type": "string",
            "format": "date-time",
            "title": "Incident Start"
          },
          "incident_end": {
            "type": "string",
            "format": "date-time",
            "title": "Incident End"
          },
          "trigger_metric": {
            "type": "string",
            "title": "Trigger Metric"
          },
          "baseline_value": {
            "type": "number",
            "title": "Baseline Value"
          },
          "observed_value": {
            "type": "number",
            "title": "Observed Value"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "ontology_version": {
            "type": "string",
            "title": "Ontology Version"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "latest_diagnosis_run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Diagnosis Run Id"
          },
          "latest_diagnosis_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Diagnosis Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "scenario_id",
          "dataset_ref",
          "baseline_start",
          "baseline_end",
          "incident_start",
          "incident_end",
          "trigger_metric",
          "baseline_value",
          "observed_value",
          "status",
          "ontology_version",
          "description",
          "created_at",
          "updated_at"
        ],
        "title": "IncidentResponse",
        "description": "GET /incidents, GET /incidents/{id} response."
      },
...
      "SimulatedIncidentCreate": {
        "properties": {
          "scenario_kind": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Scenario Kind",
            "default": "mixed_failure"
          },
          "seed": {
            "type": "integer",
            "title": "Seed",
            "default": 42
          },
          "num_intents": {
            "type": "integer",
            "maximum": 10000.0,
            "minimum": 1.0,
            "title": "Num Intents",
            "default": 500
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "type": "object",
        "title": "SimulatedIncidentCreate",
        "description": "Request that the harness materialise one demo scenario as an Incident."
      },
...
    }
  }
}